+++ docs/tr1/CHANGELOG.md
@@ -2,6 +2,7 @@
 - added `/sfx` command
 - added `/nextlevel` alias to `/endlevel` console command
 - added `/quit` alias to `/exit` console command
+- added an option to toggle the in-game UI, such as healthbars and ammo text (#1656)
 - changed the easter egg console command to pack more punch
 - changed `/set` console command to do fuzzy matching (LostArtefacts/libtrx#38)
 - fixed console caret position off by a couple of pixels (regression from 3.0)